Files importer: lower memory usage for huge directories
When a single directory has tons of images, we have to avoid putting
every file in the queue at the same time. In the extreme case, a site
could have all files in the same directory, which results in a list of
all files being stored in RAM.

This addresses that by only putting 5k files at a time in the queue to
process. If there are more than that, we add the first 5k files and a
placeholder to be processed after that. When we get to the placeholder,
we read the directory list again and add the next 5k files until all
files have been processed. This adds approximately N/5000 more disk
operations, but can significantly reduce RAM usage.

The 5000 limit was just a guess. We can adjust that if necessary.

Note: While this effectively addresses the problem of a few directories
with lots (> 50k) of files, RAM usage will likely still be pretty high
if there are many directories with lots of files each, though lower than
it would be without this fix.