wp bulk upgrade initial commit
Named `vip site upgrade` for now to avoid collision with existing site
update command. For now it just starts upgrades and lists the sites.
Eventually we need to list the related host actions in the API so that
we can easily poll them. We can also optionally take a site ID and use
the same command for single-site upgrades as well as bulk upgrades.