Always use 'fast' mode for new sites.
Normally the importer sends a HEAD request for each file and skips it if
it already exists. This is designed to make imports faster by not
transfering lots of unnecessary data. If we know none of the files
exist, it's faster to skip the 404 checks though. There's an option for
this built into the importer now `--fast`, but it doesn't really get
used.

This detects "new" sites (currently defined by less than 100 images
uploaded) and skips the 404 checks automatically.