Fix sandbox stop
This finishes transitioning sandbox stop logic to the
/sandboxes/:id/stop endpoint. This is needed because the /container/stop
endpoint isn't granular enough to give start/stop access to a specific
set of containers (i.e. on your sandbox host). It is important because
it means we're now stopping all containers on your sandbox host for a
given site when you run `vip sandbox stop`. In practice, this is unlikely
to change anything unless you were creating multiple running sandbox containers
at once.

This PR also removes the maybePrompt helper function, which we were
using to decide which sandbox container to take action on if you had
multiple. We were only using it in the `vip sandbox pause` subcommand
anymore, which is also being removed becuase there is no
`/sandboxes/:id/pause` endpoint (yet).

Finally, this PR updates some variable names from "container" to
"sandbox" because we're actually passing sandbox objects around, which
are similar to containers, but have an ID field necessary for the
/sandboxes endpoints.