Add login, logout commands
Instead of an explicit login command, we'll prompt to login any time
there's not a valid token already. The token is stored in the system's
keychain via keytar, which is a x-platform library to securely store
passwords. Since the token is a JSON Web Token, the expiry date is
embeded and we can use that to determien whether the current token is
valid or not.