Remove gnome keyring
Revert "Remove gnome keyring"

This reverts commit 14bc03331a43e8d24cca2279d5f4e5a215784f86.

Add script to test in linux

We can use docker to easily test in a linux environment

Switch to keytar-prebuild

Fallback to insecure storage

On systems with a keychain (MacOS, Windows, Ubuntu with gnome keyring), we'll
use that to securely store tokens.

On systems without a secure method of storage, we'll store tokens in plaintext
on the disk using the node-localstorage module.

In browsers, we store the token in localStorage.