Show a different message if a sync recently ran, but is done
For various reasons, we can't run another data sync immediately after a
sync finishes. If the sync lock is still in place, but the previous sync
is done, we'll explain that a sync was recently done and it will be a
few minutes rather than claiming that a sync is currently running.