Better ProxyCommand for Windows
Using ssh to tunnel traffic instead of nc is more reliable on Windows.
Eventually we'll switch everyone to this new ProxyCommand, but it's
limited to Windows for now.

This requires at least the following SSH config:

```
Host proxy.automattic.com
    User <Matticspace Username>
```

We may want to allow users to configure their matticspace username in
the CLI to avoid needing this extra config.