Port sanitize_file_name
This copies the functionality from sanitize_file_name in WordPress. We
follow our javascript coding standards and name the function
sanitizeFileName. Since we want to use this a bit differently, we also
introduce validateFileName, which just checks if the given name is
valid. Using this new function, we check the given filename and error
if it won't work on VIP Go.