Add UUID
We need UUIDs to track usage data. We don't care who the user actually
is, but we need to be able to see problematic flows.

This generates a unique ID that cannot be reversed or
otherwise used to figure out who the user is. The ID is persisted in
whatever keychain is used on that given system.