Add scripts to build config file
Integrated as part of prepare (to handle the local version of the file)
and prepublishOnly (to handle the publish version of the file).

Technically, both local and publish configs are written on publish since
we don't have any npm-specific scripts to only fo something on a local
install. However, the latter `publish` release overwrites the existing
local copy of the config.