Add $PAGER support
Adds support to use a pager for long commands. By default we use `less -FRX`
which disables the init step (so the screen doesn't completely clear
when less starts), sends the raw output to less (so colors work), and
exits if less than one full screen is output.

We also honor the $PAGER environment variable, so the program that's
used for paging can be customized per http://manpages.ubuntu.com/manpages/bionic/en/man1/man.1posix.html

An easy way to test this is to change the `vip app list` to output 30
results instead of 10.

This is a good start, but it would be better if we could page over
results and write 10 at a time. It would also be nice if that was
pausable (like stream.pause()) so we could lazy load results (or never
load results) that aren't going to be on screen anyway.

See: #116