Fix `vipgo api` with jq
In #393 we switched to outputting JSON via console.log directly, but
unfortunately that breaks jq because the JSON is invalid. Here we detect
whether the output is a terminal, and if so, use console.log to print
directly. If the output is not a terminal (i.e. piping the output to jq)
we format it correctly with JSON.stringify.