Always use readline to parse input
This ensures we always use readline to parse input and therefore
don't rely on bash to decide when input is worthy of being sent
to stdin (i.e. on return).

This effectively makes everything subshell mode, though if you run a
specific command, we return to your system shell as soon as the command
finishes.

Noting that this also removes the special "wp shell" mode as well, since the
normal mode is now able to support all input. We could add that back
if there are other reasons to have it.

The primary thing this fixes is `wp help` or other similar long running
commands that expect input over time. Previously input was only sent
when the user pressed return, but now we're able to capture all input.

Fixes #363