Don't re-print command to stdout
Since we're using readline, writing directly to the stream means we end
up sending the command to the out stream as well. To get around that,
we're adding the ability to mute the stream. This is a similar approach
to what you would use to avoid printing passwords as they're typed.