Add util function to capture multiline sql statements
Validations are currently performed line-by-line using the Node `readline` module. This adds a function to enable capturing full statements using the existing line-by-line processing. Once a pattern is matched, the function captures subsequent lines through to the semicolon ending the statement. These are returned as an array of matching statements, where each statement is an array its lines.