FORNO-1254: SQL Import: Add multisite primary domain validation (#1064)
* Add util function to capture multiline sql statements

Validations are currently performed line-by-line using the Node `readline` module. This adds a function to enable capturing full statements using the existing line-by-line processing. Once a pattern is matched, the function captures subsequent lines through to the semicolon ending the statement. These are returned as an array of matching statements, where each statement is an array its lines.

* Add validation to check primary domain is mapped

Adds a validation to confirm the primary domain specified in the SQL dump is mapped to the target environment

* Fix lint issues

* Apply replacements to the domain found in the SQL file

Previously this validation only checked if the domain found in the SQL file was mapped to the environment. This didn’t consider that the found domain may end up being replaced by the `—search-replace` flag.

* Add tests

* Fix issue with validation of single line statements

Validation was only working for statements which spanned multiple lines. This allows capturing of single-line statements as well. Adds tests for this scenario as well.

* Improve validation message

Change wording and formatting to improve clarity