Don't pipe output through less
This was originally meant to be useful but it's actually more
frustrating and a bit of an anti-pattern when working with a CLI tool.
It's especially frustrating because you can't easily disable the
behavior either. Plus, we're assuming that less or PAGER will be
available which may not always be the case.

Instead, let's just output via console.log and let the user decide
any additional piping they want to do.