Add pollUntil function
This function is used to poll another function until a certain condition is met. It takes three parameters:
`fn`: the function to poll
`interval`: the poll interval
`isDone`: the function that takes the return value of `fn` as parameter, checks if certain conditions are met, and returns a boolean. If it returns true, polling will stop.